Moved view menu listener
