+++ src/jumpin/view/game/menu/menus/solver/SolverMenu.java
@@ -13,8 +13,8 @@ import jumpin.view.constants.ComponentSize;
 import jumpin.view.constants.ViewConstants;
 import jumpin.view.factory.ComponentFactory;
 import jumpin.view.game.menu.MenuFrame;
-import jumpin.view.game.menu.listener.MenuEvent;
-import jumpin.view.game.menu.listener.MenuListener;
+import jumpin.view.listener.MenuEvent;
+import jumpin.view.listener.MenuListener;
 import jumpin.view.util.GroupBox;
 
 /**
